Initial progress on adding an update method to Session
The associated function `update()` can be called on any instance of
Session (hence, it is a method) - and all business logic to persist data
to the session (which is Redis backed) is neated handled (implementing
separation of concerns).

As a future improvement, we could look at decoupling the data store (or
backend), which for now is hardcoded around Redis.